ProjMgtService-UserStoryFeature: Refactoring the package for dtos, and added new required dependencies for transforming DTOs to Models and viceversa.
