+++ src/main/java/com/umssonline/proymgt/common/dto/UserStoryDto.java
@@ -1,0 +1,4 @@
+package com.umssonline.proymgt.common.dto;
+
+public class UserStoryDto {
+}
