--- src/main/java/com/umssonline/proymgt/controllers/request/UserStoryDto.java
@@ -1,4 +1,0 @@
-package com.umssonline.proymgt.controllers.request;
-
-public class UserStoryDto {
-}
